# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Conversation:
    properties:
      conversation_id:
        type: integer
      last_message_at:
        type: string
      partner_acc_id:
        type: integer
    type: object
...
  models.InboundMessage:
    properties:
      conversation_id:
        description: For sendMessage, loadHistory, markSeenMessage
        type: integer
      limit:
        description: For loadHistory
        type: integer
      offset:
        description: For loadHistory
        type: integer
      partner_acc_id:
        description: For sendMessage, markSeenMessage
        type: integer
      read_time:
        description: For markSeenMessage
        type: string
      text:
        description: For sendMessage
        type: string
      type:
        description: e.g., "sendMessage", "loadHistory", "markSeenMessage"
        type: string
    type: object
...
  models.Message:
    properties:
      content:
        type: string
      conversation_id:
        type: integer
      is_seen:
        type: boolean
      message_id:
        type: integer
      sender_acc_id:
        type: integer
      sent_at:
        type: string
    type: object
...
  models.OutboundMessage:
    properties:
      conversation_id:
        description: for "newMessage", "seenMessage", "messageHistory"
        type: integer
      conversations:
        description: For "conversationList"
        items:
          $ref: '#/definitions/models.Conversation'
        type: array
      id:
        description: For "yourID"
        type: string
      message:
        allOf:
        - $ref: '#/definitions/models.Message'
        description: For "newMessage"
      messages:
        description: For "messageHistory"
        items:
          $ref: '#/definitions/models.Message'
        type: array
      read_time:
        description: For "seenMessage"
        type: string
      text:
        description: For "error"
        type: string
      type:
        description: e.g., "yourID", "newMessage", "messageHistory", "conversationList",  "error"
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for medical records management app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: MeReMa Server API
  version: "1.0"
paths:
...
  /v1/communication/chat:
    get:
      consumes:
      - application/json
      description: Establish a WebSocket connection for real-time chat between doctor
        and patient.
      parameters:
      - description: Comm request (type = {'sendMessage', 'loadHistory', 'markSeenMessage'})
        in: body
        name: rqeuest
        required: true
        schema:
          $ref: '#/definitions/models.InboundMessage'
      produces:
      - application/json
      responses:
        "101":
          description: type = {'yourID', 'newMessage', 'messageHistory', 'conversationList',
            'error'}
          schema:
            $ref: '#/definitions/models.OutboundMessage'
      security:
      - BearerAuth: []
      summary: WebSocket connection for chat (doctor, patient)
      tags:
      - communications
securityDefinitions:
  BearerAuth:
    description: Provide the JWT token as a header with format "Bearer \<token\>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"