# models/comm_models.go
package models

import "time" // For mock repository structs

// --- Mock Repository Structs (to avoid direct dependency on actual repository package) ---
// These would normally be in your repository package. We define simplified versions here.
type Conversation struct {
	ConversationID int64      `json:"conversation_id"`
	PartnerAccID   int64      `json:"partner_acc_id"`
	LastMessageAt  *time.Time `json:"last_message_at"`
}

type NewMessage struct {
	ConversationID int64  `json:"conversation_id"`
	Content        string `json:"content"`
}

type Message struct {
	NewMessage
	MessageID   int64     `json:"message_id"`
	SenderAccID int64     `json:"sender_acc_id"`
	IsSeen      bool      `json:"is_seen"`
	SentAt      time.Time `json:"sent_at"`
}

type InboundMessage struct {
	Type           string     `json:"type"`                      // e.g., "sendMessage", "loadHistory", "markSeenMessage"
	ConversationID int64      `json:"conversation_id,omitempty"` // For sendMessage, loadHistory, markSeenMessage
	PartnerAccID   int64      `json:"partner_acc_id,omitempty"`  // For sendMessage, markSeenMessage
	Text           string     `json:"text,omitempty"`            // For sendMessage
	Limit          int        `json:"limit,omitempty"`           // For loadHistory
	Offset         int        `json:"offset,omitempty"`          // For loadHistory
	ReadTime       *time.Time `json:"read_time,omitempty"`       // For markSeenMessage
}

// OutboundMessage represents messages going FROM the server TO the client.
type OutboundMessage struct {
	Type           string         `json:"type"`                      // e.g., "yourID", "newMessage", "messageHistory", "conversationList",  "error"
	ID             string         `json:"id,omitempty"`              // For "yourID"
	Message        *Message       `json:"message,omitempty"`         // For "newMessage"
	Messages       []Message      `json:"messages,omitempty"`        // For "messageHistory"
	Conversations  []Conversation `json:"conversations,omitempty"`   // For "conversationList"
	Text           string         `json:"text,omitempty"`            // For "error"
	ConversationID int64          `json:"conversation_id,omitempty"` // for "newMessage", "seenMessage", "messageHistory"
	ReadTime       *time.Time     `json:"read_time,omitempty"`       // For "seenMessage"
}

// --- End Mock Repository Structs ---